feat: Use TUE_ROS_VERSION to select ros version
This commit introduces the use of TUE_ROS_VERSION for selection of the ros target to be installed.

Add warnings stating that setting this variable in the tue-env workspace will be made mandatory in the future.
The default value of TUE_ROS_VERSION is set to 1.

Warnings about TUE_ROS_VERSION have been set to once per run of tue-get commands