ci: Add build for ROS 2 in Dockerfile
Added ROS_VERSION as input argument to the Dockerfile, which can either be the integer 1 or 2 and nothing else,
similar to how the environment variable is defined after installing ROS and sourcing, with ROS 1 being default.
For ROS 2 the target ros2 is used instead of ros.

An alias target has been added to the targets repository for the target ros.
This allows the use of environment variable `ROS_VERSION` in the target name

The bootstrap script uses the new ros-version argument for Ubuntu 20.04

Co-authored-by: Elwin van der Hall <e.hall@avular.com>